tests: Extract node related helpers to a dedicated file
`utils` is overloaded with multiple functionalities which makes it hard
to properly maintain its content.

This change takes all node related code to a dedicated file, still under
the same `tests` package.
